requirements.txt
aiodns==3.0.0
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
Babel==2.9.1
ccxt==2.8.84
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.2
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
numpy==1.24.2
packaging==23.0
pandas==1.5.3
patsy==0.5.3
pycares==4.3.0
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
scipy==1.10.1
six==1.16.0
statsmodels==0.13.5
urllib3==1.26.14
yarl==1.8.2
